+++ src/main/java/edu/mum/cs/cs544/spring/hvp/data/domain/Task.java
@@ -21,7 +21,7 @@ import javax.persistence.TemporalType;
 @Table(name = "task", catalog = "hvp")
 public class Task implements java.io.Serializable {
 	private static final long serialVersionUID = 1089077518414238405L;
-	private int id;
+	private Long  id;
 	private String name;
 	private short status;
 	private String description;
@@ -34,13 +34,13 @@ public class Task implements java.io.Serializable {
 	public Task() {
 	}
 
-	public Task(int id, String name, short status) {
+	public Task(Long  id, String name, short status) {
 		this.id = id;
 		this.name = name;
 		this.status = status;
 	}
 
-	public Task(int id, String name, short status, String description, Date startDate, Date endDate) {
+	public Task(Long  id, String name, short status, String description, Date startDate, Date endDate) {
 		this.id = id;
 		this.name = name;
 		this.status = status;
@@ -60,11 +60,11 @@ public class Task implements java.io.Serializable {
 	@Id
 	@GeneratedValue
 	@Column(name = "id", unique = true, nullable = false)
-	public int getId() {
+	public Long  getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long  id) {
 		this.id = id;
 	}
